Improve typing and no inherit options
