+++ src/index.ts
@@ -2,11 +2,13 @@ import type MarkdownIt from "markdown-it/lib"
 import { admonitions } from "./directives/admonitions"
 import { code } from "./directives/code"
 import { images } from "./directives/images"
-import directivePlugin, { IOptions as IDirectiveOptions } from "./directives/plugin"
+import directivePlugin from "./directives/plugin"
+import type { IOptions as IDirectiveOptions } from "./directives/types"
 import { tables } from "./directives/tables"
 import { roles } from "./roles/main"
 import { html } from "./roles/html"
-import rolePlugin, { IOptions as IRoleOptions } from "./roles/plugin"
+import rolePlugin from "./roles/plugin"
+import type { IOptions as IRoleOptions } from "./roles/types"
 import { math } from "./roles/math"
 
 /** Allowed options for docutils plugin */
