+++ src/roles/math.ts
@@ -1,6 +1,7 @@
-import MarkdownIt from "markdown-it"
+import type MarkdownIt from "markdown-it/lib"
 import type Token from "markdown-it/lib/token"
 import { IRoleData, Role } from "./main"
+import { IOptions } from "./types"
 
 const INLINE_MATH_RULE = "math_inline"
 
@@ -14,18 +15,13 @@ export class Math extends Role {
   }
 }
 
-export interface IOptions {
-  roles?: Record<string, typeof Role>
-  mathRenderer?: (content: string, opts: { displayMode: boolean }) => string
-}
-
 export function inlineMathRenderer(md: MarkdownIt, options?: IOptions): void {
   // Only create the renderer if it does not exist
   // For example, this may be defined in markdown-it-dollarmath
   if (!options?.roles?.math || md.renderer.rules[INLINE_MATH_RULE]) return
 
   md.renderer.rules[INLINE_MATH_RULE] = (tokens, idx) => {
-    const renderer = options?.mathRenderer ?? (c => md.utils.escapeHtml(c))
+    const renderer = options?.opts?.math?.renderer ?? (c => md.utils.escapeHtml(c))
     const token = tokens[idx]
     const content = token.content.trim()
     const math = renderer(content, { displayMode: false })
